Avoid circular import by moving convert_to_oe
